// src/proto/dapr/proto/durabletask/v1/orchestration_pb.d.ts
export declare type TaskFailureDetails = Message<"TaskFailureDetails"> & {
  /**
   * @generated from field: string errorType = 1;
   */
  errorType: string;

  /**
   * @generated from field: string errorMessage = 2;
   */
  errorMessage: string;

  /**
   * @generated from field: google.protobuf.StringValue stackTrace = 3;
   */
  stackTrace?: string | undefined;

  /**
   * @generated from field: TaskFailureDetails innerFailure = 4;
   */
  innerFailure?: TaskFailureDetails | undefined;

  /**
   * @generated from field: bool isNonRetriable = 5;
   */
  isNonRetriable: boolean;
};